Refactor code to comply with Checkstyle standards
Currently, the code does not fully adhere to the coding standards enforced by Checkstyle,
which has led to violations during code quality checks.

This commit refactors the code to ensure it complies with the Checkstyle configuration.
Changes include adjustments to import order, line breaks, spacing, as well as the addition of necessary Javadoc comments.

Ensuring compliance with Checkstyle helps maintain a consistent coding style across the project and improves readability and maintainability.
