+++ src/main/java/elara/command/FindCommand.java
@@ -1,9 +1,9 @@
 package elara.command;
 
 import elara.storage.Storage;
+import elara.task.InvalidInputException;
 import elara.task.TaskList;
 import elara.ui.Ui;
-import elara.task.InvalidInputException;
 
 /**
  * Represents a command that finds tasks containing a specified keyword in the task description.
