+++ src/main/java/elara/command/ListCommand.java
@@ -1,7 +1,6 @@
 package elara.command;
 
 import elara.storage.Storage;
-import elara.task.InvalidInputException;
 import elara.task.TaskList;
 import elara.ui.Ui;
 
@@ -10,7 +9,7 @@ import elara.ui.Ui;
  */
 public class ListCommand implements Command {
     @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws InvalidInputException {
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
         ui.listTasks(taskList);
     }
 }
