+++ src/test/java/elara/task/ToDoTest.java
@@ -1,8 +1,7 @@
 package elara.task;
 
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.api.Test;
 
 /**
  * Unit tests for the ToDoTask class.
